[Kernels] Fixes the construction of static tuples with a list (#6796)
This fixes the construction of static tuples using a list. We do not
have a concept of an `if constexpr` so the prior implementation hits a
few errors. Along the way, this simplifies how we construct a kgen.list
by passing parameters as regular parameters (not meta-parameters). This
allows us to overload the function and not do manual mangling.

Resolves issue #6786

BUFFER_ORIG_REV_ID: a82c3e1025d4c9c46c6b55a4d98a60cee06d0747