[Kernels] Add an optimized memcpy implementation (#7177)
Memcpy is useful enough (used by broadcast) and others that it is
worthwhile to provide an optimized implementation of it. This implements
the same strategy outline in
https://github.com/nadavrot/memset_benchmark and
https://msrc-blog.microsoft.com/2021/01/11/building-faster-amd64-memset-routines/.
The difference is that we use our vectorize generator for the last step
instead of using the manual vectorization loop that is used in the
reference implementation.

In the future, we should search for the decision tree like we did in the
prior milestone. Once we have search, then this can be delegated as a
reference implementation.

BUFFER_ORIG_REV_ID: c3f746fb7ced684d324e144f810668bbd979a736