[KGEN/LIT/Kernels] Move KGEN structs to LIT and fix POP<->KGEN dep (#7189)
This PR fixes a longstanding major layering problem with the KGEN
dialect secretly depending on the POP dialect due to its use of
`POP::PointerType`. This PR resolves this by moving all the struct ops
to the LIT dialect (taking the offending StructGEPOp along with it).
Moving structs to LIT also mean that we can formally drop support for
handling structs in certain parts of the post-LIT pipeline.

Unfortunately, the remaining uses of `POP::PointerType` in KGEN are not
trivial to abstract away, but they're not as important as StructGEPOp,
so I dropped them (creating some verifier holes...).

Fixes #5742

BENCHMARK_ORIG_REV_ID: 8695fef8c94eb6ca7221efbaf1161e03d22a98f4