[Kernels] Move softmax to use the new high-level vectorize generator (#7269)
This moves the 3-step softmax kernel to use the high-level vectorize
generator. Care is taken to ensure we are still applying the stategy
outlined by Ian in [Internal link]

There is a trick here to get "simd accumulator in step2" of the kernel
while still using the vectorize kernel. We introduce a function which
accumulates into a scalar or a vector value depending on the current
simd size.

NN_ORIG_REV_ID: 36407b579c4cfc2ae115a5d8d66723f8e3108705