[Kernels] Introduce a new repeat function (#7287)
The repeat function takes a count and another function and repeatadly
calls the function until the count is reached. It passes the current
iteration index into the function via a parameter.

There were a bunch of places where we use this pattern (corresponding to
compile-time unrolling) because we need an index constant. This just
reimplements this in two places.

BUFFER_ORIG_REV_ID: aa624c345e0b366c18e547dcb3becf64fe248c83