[Kernels] Move step2 of the 2-step softmax to use vectorize (#7276)
This moves step2 of the 2-step softmax to use vectorize. This removes a
bunch of code which is now handled by the vectorize generator. The only
thing left is to move the first step of the 2-step softmax to use
vectorize.

NN_ORIG_REV_ID: bcfa4bfd42f4814a74b102c1ddd14ef6a2194d6f