[Kernels] Make the unary element-wise operations run in parallel (#7831)
This makes all the exported unary operations run in parallel. The trick
is to create the runtime via lit within the gtest (via an exported
function) since a runtime cannot be shared across dylib boundaries. This
however creates another intersting issue where the recently committed
parallel matmul was expecting a runtime created in the gtest rather than
in compiler-rt. We disable the parallel matmul for now and revert to a
sequential version. We were planning on sinking the parallel dispatch of
matmul for C++ land to lit once we figure out how to run parallel
kernels in lit anyways.

Thanks to @arjunsurendran24 / @hengjiew for a lot of help.

ALGORITHM_ORIG_REV_ID: 67bde41ef5c8daa330bcab9965aaef963b3cc52f