[Kernels] Export the mean kernel that works on 1D and 2D ranked tensors (#7873)
This exports the mean kernel which works on both 1D and 2D ranked
tensors. For the 1D case, the output would be a scalar value (stored in an output
buffer which has one rank less than the input). For the 2D case, we
require the axis to be 1 and that would compute the mean for each batch.

BUFFER_ORIG_REV_ID: 0dded97d932aabb08bd9ad48746bb5dee5a64f7f