[LIT][stdlib] Rework the coroutine implementation (#7993)
This PR reworks the coroutine implementation to abstract the LLCL
runtime entirely out of the implementation details in both LIT and KGEN
(it was already abstracted out of KGEN). This is done by changing the
meaning of the coroutine "context" to be a completion callback closure.
When a coroutine returns, it invokes this callback, and that callback is
customizable by client code.

This enables a default `Coroutine[T]` implementation in the standard
library that has no relation to LLCL but which is functional on its own.
The LLCL `Future` type has been renamed more appropriately to `Task` and
it works more or less the same by overwriting the completion callback of
the coroutine handle to setting an async chain ref to the avaiable state.

This enables a more efficient TaskGroup implementation whereby the completion
callback of task group tasks is set to decrementing the TaskGroup atomic
counter directly.

Note that the lower-level implementation of POP coroutine operations has
barely changed. This is a good sign that those operations represent a
good base abstraction on top of which LIT async functions are
implemented.

The most important aspect of this change is that the `!pop.coroutine`
returned when invoking an async function does nothing on its own. It
__must__ be wrapped in a `Coroutine` library type. When LIT has
literals, this conversion will be automatic.

Fixes [Internal link]
Fixes [Internal link]

RUNTIME_ORIG_REV_ID: 675891687c0df267d58daa1b5d3e12f4d1b6d2f9