[KGEN] Change coroutine lowering to async (#9371)
The async coroutine lowering is more efficient in many ways. First, we
can finally get rid of the spin loop in `Coroutine.__del__` which was
the bane of my existence. Second, the coroutine splitting on async
coroutines is far simpler, which means we have more control over the
generated code. This allows us to tackle
[Internal link] because the allocas
are guaranteed to be placed at the beginning of the resume function.

RUNTIME_ORIG_REV_ID: 1fb5aef7ae55bfa6c9eeec6ab4db85b8b6a93f98