[Stdlib] Switch the gather, reduce, and mandelbrot to use parallelize (#10250)
This switches the gather, reduce, and mandelbrot function to use the
parallelize function inplace of the parallelForEachN function. This is done as
a first step to remove the parallelForEachN function from the stdlib and
switching these functions to use generators.

We also inline more methods in the stdlib to reduce the number of
function calls.

Note: We had to disable the pretch the gather functions
since otherwise LLVM complains with

```
immarg operand has non-immediate parameter
  %162 = load i32, ptr %161, align 4
  call void @llvm.prefetch.p0(ptr %160, i32 %162, i32 %164, i32 %166)
immarg operand has non-immediate parameter
  %161 = load i32, ptr %160, align 4
  call void @llvm.prefetch.p0(ptr %159, i32 %161, i32 %163, i32 %165)
```

This is weird because the prefetch is not using the immarg operand as
they are passed as a constant.

KERNEL_BENCHMARKS_ORIG_REV_ID: e94d1c9c20ba1ffec68e34a3e8933681b6081498