[Stdlib] Add initial element-wise generator (#10356)
This adds an initial element-wise generator that can be used to replace
most of the code in element-wise. The generator will take an compute
function and would vectorize, unroll, and parallelize over the
function. The function take a rank-sized index variable,
but currently only the first index is used (this will be generalized
in subsequent PRs).

All the element-wise functions are now generated using this generator.

As mentioned this will be generalized and there are some things that
need to be figured out, but at least we can start playing with it to
figure out the best interface and semantics.

Part of [Internal link]

ALGORITHM_ORIG_REV_ID: 7375aae9696141908308ffd6ae933cfeac795c61