[mojo-lang] Switch borrowed convention on by default! (#10707)
Arguments to functions now default to the 'borrowed' convention.
This is equivalent to passing things by "const&" in C++ with two
exceptions:

1) A future borrow checker will make sure there are no mutable
   aliases with them.
2) register_passable values are passed directly instead of using
   an extra reference wrapper, and requiring complexity to avoid
   it.

Achieving this required fixing a few remaining bugs in the compiler,
and updating a ton of testcases.  We should discuss changing the
kgen default to 'borrow' because it is so pervasive and requires
people using !kgen.signature directly to know about it.

ALGORITHM_ORIG_REV_ID: f4014a0621dca8a4cee82098cc85bd6c5c56c633