[Mojo] Initial renames and Reorg of the Stdlib (#11682)
This does a first pass on renaming lit to mojo and reorganizing the
files. It does not move methods within the files (that's left for a
subsequent PR). This is also by no means the final placement of the
files, since some of these are ambigous as to what category they belong
in (e.g. slice is now placed in NN).

We also needed to make modifications to KGEN to allow us to recursively
search through the auto-import directory.

MOGG_ORIG_REV_ID: 038c92586c0d27ae2233a1715928dadf99777310