[mojo-lang] Changes for memory-only `__init__` method support. (#12188)
This revises how `__init__` is defined for memory only types.  It
specifically removes support for the `->Self` syntax (but keeps it
just for `@register_passable` types), and changes the internal
modeling of the self parameter to use a new 'self_init' convention
instead of hacking the byref_result convention.

This requires a lot of far flung work that fixed other bugs in the
stack, including we had broken parameter inference for this case
(which is now fixed) and we can now get rid of the
'isInitFnWithByRefResultSelf' bit on ASTDecl.

This fixes #12042 and makes room for more dtor work.

BENCHMARK_ORIG_REV_ID: fd52d7f04f524471d4e85d3dc54b9a79bf0fc706