[Stdlib] Add a slice struct to allow slicing arrays (#12292)
This adds a slice struct (lower case on purpose) which is going ot be
used for the demo. Ideally we want to place this struct in Stdlib/Slice
but that will conflict with NN/Slice. This is placed in NN/Slice until
we come up with a better name for NN/Slice .

NN_ORIG_REV_ID: faeb9b4a6a23dfec8aa20d43101cd8e06d61b149