[Profiling] Cleanup kLLCL and kMojo profiling entries (#12349)
This gets us to a state where
`-DMODULAR_LLCL_MAX_PROFILING_LEVEL=01111`, or `cmake-modular-release
--build-profile=profiling_default` gives us a readable profile with all
the 'important' things captured, but without overwhelming detail or
confusing entries (at least in my eyes).

- Retire all the redundant mop.kernel.* entries. MO primitives without a
Mojo impl (ie just the not-long-for-this-world shape functions) are
still profiled on the C++ side, but I move them to the kMojo level 1 and
use "mojo.*" for consistency.
- kLLCL level 1 includes only 'important' entries. I'm including thread
sleeping in those since once a worker sleeps we're at the whim of the OS
for it to come back promptly so we should be watching for that. A few
other concerning situations also included here, such as needing to run
an addTask closure immediately because the task queue is full.
Basically, if you see 'llcl.*' in the profile it is worth looking at.
- But also add a kLLCL level 3 for tracking *every task and waiter
execution*. Can turn that on when there are unexplained blank spots in
the profile to try to track down when Mojo kernels have forgotten to
trace, or at least distinguish work being done on our side from untraced
primitives (eg free). Stephen and I needed to do this by hand the other
day, this makes it just a kLLCL level change.
- Stop the profiling entries for *async* Mojo kernels when they return
to MEF, not when the last of their launched sub-tasks completes. I think
these are confusing the heck out of us, and the automatically added
'mojo.*.task' entries should be good enough to track when and where the
true kernel work is being done. It was a bit of a dance to also support
the sync kernels.
 - Add missing profiling for the Mojo matmul pack.
- Distinguish setup vs warmup vs run phases in ExecutionDriver.cpp
executeModel.
- Have just one Mojo profiling entry for jitted kernels, but capture the
kernel name from the C++ side and details on the Mojo side.
- Update Profiling.h documentation for the Trace::Type enum entries so
clear what is profiled by which flavor and level.

MOGG_ORIG_REV_ID: 1244cda4ac7877c1423d8cf2d0bd7b11e725f4e7