[mojo-lang] List literals emit TupleLiteral (#12576)
Python lists are heterogeneous, which make them structurally identical
to
TupleLiteral. In time we will get a different ListLiteral type that is a
heterogenous type, but this gets lists off the ground for now.

The main change is that `_type:` specifier on operations is no longer a
list, but it is `_type: None` for no result type and `_type: Int` for a
single result type (multiple results are not supported anyways).

ALGORITHM_ORIG_REV_ID: ef8f4a4fc9ad9c114698f650a5ee54d953ad59ea