[Kernels] Report DType and rank errors from the Kernels (#12685)
The kernels know exactly what types / ranks are supported, so we can do
the reporting from the kernels. This means that a future commit can
remove the dtype / rank checks from the C++ shims and we can have a
single source of truth for what is and is not supported.

BUFFER_ORIG_REV_ID: 591382a7783749ec98f3e8a7d01d84a956c6736d