[Kernels] Implement a self attention block in the Kernels (#13147)
This lays the ground work for interesting fusion with matmul. It shows
how to implement a fused attention block using KGEN. The goal of this
not to be efficient (that will happen in subsequent PRs), but to just to
lay down the API.

For now, this only implements a single-head attention (also called
scaled dot product). A Scaled Dot Product is a primitive operation in a
transformer which leverages both the Batched Softmax along with Matmul.
The pseudocode of the naive implementation is:

```
def ScaledDotProduct(Q, K, V):
  return dot(Softmax(dot(Q, tr(K))/sqrt(d), V)
```

were dot(A, B) is a matrix multiplication, tr(A) is the transpose of A.
The sqrt(d) is called the temperature of the scaled dot product, and is
as a regularizer to ensure that the variance of matmul(Q, tr(K)) in 1.

There are other things that we can do to optimize this kernel even more:

- actually perform the fusion.
- remove the temporary allocations (both within and outside the kernel).
We currently allocate a tensor because it will always be correct. We can
be smart though and reuse the output tensor.

We also need to add more test cases.

BUFFER_ORIG_REV_ID: 7ac8f721f9be6b05b247622bd7ffbef7fd3c4fa9