[mojo-stdlib] Un-revert printable `StaticIntTuple` (#13275)
Allow for a `StaticIntTuple` to be printed in a variadic sequence of
other printable standard library type objects.

Also, don't print an extra space in one-element tuples like `(1,)`.
That's how Python does it: `x = (1, ); print(x)` outputs `(1,)`.

This was reverted because it was stacked on top of a commit that caused
leaks, not because it caused leaks itself.

NN_ORIG_REV_ID: f3c9ec6b6f5f6e84a155ede9a72a48d82da47d72