[Kernels] Flush denormals to zero for all kernels (#13542)
Based on Srini's suggestion, there is a huge perf benifit from flushing
denormals to zero for kernels. This is also a common pattern found in
both
[ort](https://github.com/microsoft/onnxruntime/blob/926ae7d7862addd544b5273c75a0fa0895ab54ee/onnxruntime/core/common/denormal.cc#L31),
[pt](https://github.com/intel/intel-extension-for-pytorch/blob/4c1882e4d0c5d65a629ad7341cd7e35739f494dd/csrc/cpu/tpp/init.cpp#L43),
as well as
[tf](https://github.com/tensorflow/tensorflow/blob/master/tensorflow/tsl/platform/threadpool.cc#L65)
it is also used by other libraries such as ruy.

We still need to double check this version of the implementation, but
acording to prior analysis which was done in C++ land this gives a 2x
speedup on gpt2.

See https://en.wikipedia.org/wiki/Denormal_number and
http://stackoverflow.com/questions/9314534/why-does-changing-0-1f-to-0-slow-down-performance-by-10x

ALGORITHM_ORIG_REV_ID: 7b8663aca7c9feffe006e34bbcd57933f2f9ad0f