[Stdlib] Add matmul_parallel_async to stdlib (#15582)
Prior to this PR there was no parallel matmul API in the LinAlg lib and
the multithreading logic was scattered around the MatmulKernels.mojo
shim.

The API uses parameters over args where possible so that it can be
jit'ed efficiently by MOGG. The caller is expected to dispatch from
runtime args to parameters, although different overloads could be
provided in the future.

The shims in MatmulKernels.mojo are refactored to use the new API so it
is covered by existing tests.

LINALG_ORIG_REV_ID: 826cf4476a0b4ef032ff26b132f4012bcbdcfd85