[MOGG] Expose hooks to block kernel aysnc-ness (#15716)
Until we have the full 3 stage kernel/elaboration support where the
async nature of the kernel falls out of elaboration we need a way to
force the behaviour in the graph compiler. This adds a flag to the hooks
for elementwise and gather which allows them to be forced into a simple
synchronous kernel.

This enables them to be soft fuse as per the design in
[Internal link]

Longer term this is something which would fall out of elaboration with
compiler introspection stages

ALGORITHM_ORIG_REV_ID: 20775c8f958e575be010f0775f9a9fd11654838d