[Stdlib] Generalize function mapping flat index to ND indices (#15926)
The previous function extracted the starting indices of the nth row in a shape.
This can be generalized to getting the starting indices of the nth subvolume of
a shape, which is useful in other places in the kernel library.

ALGORITHM_ORIG_REV_ID: c316117dee4c3866989725d8c12fd13bc2bb66c6