[Kernels] Perform a "small-matmul" optimization for batched matmul (#16192)
One of the customer models has a small batched matmul (that practically
should be expressed as a reduction instead) so perform the same bypass
that we did for matmul to not call the bmm optimized for large sizes.

I do not care about duplication at this point, since this all needs to
be redone soon.

MOGG_ORIG_REV_ID: 077f774ed14e8821d163c85152dc9a566a9ec28e